Changed mask drawing function
Previously using a function from the PIL library to draw the mask, and this was combined with a fineGrid step so pixels are in the mask if at least 50% of their area is inside the polygon.

Recommendation from IBSI is that use the simpler method where if the centre of the voxel is inside the polygon then the mask is true for that pixel.

Function draw.polygon from skimage toolbox is now used.

Added flags to thumbnail function to control whether to draw contours and mask boundary