Include all diagonistics info
Sometimes the overall feature extraction uses more than one configuration file (Params.yaml).  Previously the code was set to include only one set of 'diagnostics' section in the output on the assumption that any differences are not relevant, but this is not a good idea.
Change here is to include the diagnostics fields for each sub-extraction.